requirements.txt
altgraph==0.17
auto-py-to-exe==2.9.0
autopep8==1.5.7
bottle==0.12.19
bottle-websocket==0.2.9
Eel==0.12.4
future==0.18.2
gevent==21.1.2
gevent-websocket==0.10.1
greenlet==1.1.0
importlib-metadata==4.3.1
numpy==1.19.5
opencv-python==4.5.2.52
pkg-resources==0.0.0
pycodestyle==2.7.0
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.1
PyYAML==5.4.1
toml==0.10.2
typing-extensions==3.10.0.0
whichcraft==0.6.1
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.4.0
